requirements.txt
# PolyAI — PCOS Early Detection System
# Python >= 3.10
# Install: pip install -r requirements.txt

# ── Phase 1: ML ──────────────────────────────────────────────────────────────
pandas>=2.0
openpyxl>=3.1          # required to read .xlsx files
numpy>=1.24
scikit-learn>=1.3
scipy>=1.11
xgboost>=2.0
shap>=0.44
kaggle>=1.6

# ── Visualisation ─────────────────────────────────────────────────────────────
plotly>=5.18
matplotlib>=3.7
seaborn>=0.13
kaleido>=0.2          # plotly static export

# ── Phase 2: Backend ──────────────────────────────────────────────────────────
fastapi>=0.110
uvicorn[standard]>=0.29
httpx>=0.27
pydantic>=2.6

# ── Phase 2: LLM / Vector DB ──────────────────────────────────────────────────
ollama>=0.1
chromadb>=0.5
langchain>=0.2
langchain-community>=0.2

# ── Phase 2: Database ─────────────────────────────────────────────────────────
supabase>=2.4
python-dotenv>=1.0

# ── Phase 3: Frontend ─────────────────────────────────────────────────────────
shiny>=0.10
shinyswatch>=0.6

# ── Notebooks ─────────────────────────────────────────────────────────────────
jupyter>=1.0
ipykernel>=6.29
ipywidgets>=8.1
nbformat>=5.9
